Module check_graphite::assertions

Structs

Assertion
The primary property of

Enums

AssertionState
Current state of parsing the assertion

NegOp

ParseError

PointAssertion

Functions

operator_string_to_func
Take an operator and a value and return a function that can be used in a filter to return only values that do not satisfy the operator.

parse_assertion

parse_ratio
convert "all" -> 1, "at least 70% (points|series)" -> 0.7