Module check_graphite::graphite

Interact with graphite

This module defines functions that can fetch data from graphite, and the types that they return.

The two top-level items in here

Structs

DataPoint
One of the datapoints that graphite has returned.

FilteredGraphiteData
Represent the data that we received after some filtering operation

GraphiteData
All the data for one fully-resolved target

GraphiteIterator

GraphiteResponse
The result of fetch_data

Enums

GraphiteError

Functions

fetch_data
Load data from graphite

get_graphite
Fetch data from graphite